requirements.txt
discord.py>=2.4,<3.0
SQLAlchemy>=2.0,<3.0
PyYAML>=6.0,<7.0
requests>=2.32,<3.0
